Used JTAG debug to check if I'm actually reading correctly when sending commands to read Intan reg40-44 in order.
It looks good!! Setting used is TFSDIV=3, 17 bits data length, DITFS,
late frame sync.

One thing to realize is that framesync might not work properly for the
first three cycles -- i.e. after SPORTs are enabled, the first two
results received are crap. The third one corresponds to the first
command sent to Intan..2 out of the 4 results are simply 0's. This is
not fixed by checking for the SPORT_STAT for both ports in wait_samples.

The solution is to send the first command twice, then proceed normally.
Need to remember this!