intan_setup_test polls the fixed same channel from all amps. Saves 1 period of 6.4kHz (156 samples) in memory. After dumping memory into memdump1.txt (or w/e else set in plot_memdump.py), can plot the signal in mV using plot_memdump.py.
Problems...amplitude from plotting always less than 11mV...average for
the other amps seemed to be between 7-8mV. Doesn't make sense...maybe
because it's measured against the reference electrodes? That means the
ground electrode should be clipped to the ref electrode...need to try
that next.

Problems...my adjusted amp1/amp2, and amp3/amp4 are flipped? Why is this
happening?...