intan_setup_test2.asm to record 1.5 periods of 6.4kHz wave from all 32 channels of a fixed amp.
plot_memdump to handle the resulting memory output (redirected logging,
intel hex file is a bitch to read).

Problems: I'm only applying sine wave to one of the channel, but I'm not
seeing anything that makes sense...is this cross talk between channels?
Because when I record from only a fixed channel I see clearly a sine
wave...if it IS crosstalk, then I need to increase the _CS high
time..that would mean reducing the sampling freq :(

Or is it a problem of me using CONVERT63 command? Check this first...