Editted the official firmware.asm file.
To make it easier to compare, I'm taking the 16-bit ADC result and down
shifting it to 12-bits, thus allowing me to use Tim's signal chain code
as is.

Intan-setup include the DSP-offset. However, this doesn't seem to
mitigate the noise problem I see on the board. Need to figure out if
this is due to a more noisy board layout, or software problem
(saturation problem perhaps??).