The full version of firmware gives crappy results that doesn't make sense even when intan_setup_test gives great results and little cross talk when sine wave is applied.
In current firmware1.asm, I go from simply acquiring samples and saving
all channels of amplifier 1 into the WF_BUF, to running more parts of
the signal chain while saving. This has narrowed the problem down into a
matter of signal-chain length -- reading and sending commands not in
time causes the samples converted and read to not of constant rate.

As of this version, the maximum number of instructions we can execute to
still preserve signal integrity is the 8th SAA instruction of template
A.

Need to think about reducing this critical length.