Did two more test:
1. Saved the samples corresponding to amplifier 3, which is the last one
read. For firmware2 (one with no signal chain), this decreased the
number of nops from 156 to 154. For firmware1 (one with signal chain and
nops fill in), this did not change the number of nops.

This seems to suggest order of acquistion is not too much of a problem.

2. Added in an extra CONVERT_NEXT_CHANNEL at the end of radio_bidi_asm
before starting the threads, this one is not followed by wait_samples.
The purpose was to see if buffering the SPORT FIFO helps with the
problem.

It doesn't, the number of nops (only tried this in firmware2) did not
change.