Editted the Makefile for both gtkclient and headstage2_firmware to make different compilation easier by setting variable in Makefile, and use corresponding directives in the source files.
radio_basic.asm sends just the samples (or whatever preset in W1) to
gtkclient for display.

Note that gtkclient's original way of reading samples (in sock_thread()
) converts a sample into a char. This apparently by default is signed
such that 0xff+128.f = 127.0, regardless of conversion of 0xff to float
or not.

Not sure if this is intended behavior yet, but this can be solved by
making samp an unsigned char.

For some reason, the gtkclient waveform 3 now is always screwed up..