radio_AGC still doesnt work.
The algorithm itself seems legit...at least in terms of updating the AGC
gain. However, I'm not yet sure how to interpret the scaled-sample.

In Tim's version, the gtkclient interprets the received value as a CHAR
rather than as an unsigned char. This works for his results, but it
looks a bit weird, and his math would map 0x0 and 0xff all to the same
value - 0 in the plot.

I think the correct way to plot samples in gtkclient is to map 0x0 and
0xff to between +/-1. This is what Tim claims to do, but it doesn't work
that way unless the received value is read as an UNSIGNED CHAR.

This way, I can successfully plot out the raw samples from the Intan,
and be able to see amplitude changes due to the signal generator. But,
the AGC output with this setup simply oscillates from rail-to-rail (on
the plot). By reverting back to Tim's CHAR type, this no longer happens,
but my output looks non-sinusoid and the raw Intan output because
rail-to-rail oscillations...I suspect this is intimately related to how
I should actually interpret the results.

-----------------
To investigate how Tim's AGC algorithm actually operates, I need to step
through all the Blackfin calculations...very tedious.

fixedPointConvert.py: contains functions that converts binary to hex to
different fixedPoint..Q1.31, Q1.15, signedInt, Q8.8, etc.

AGC_sim.py: Emulate the entire AGC algorithm as would be executed by
blackfin.

This has been tested with 7 examples...all intermediate values checked
by stepping through and checking the register values at each step on the
blackfin using JTAG.