I think I finally understand how the AGC code works and why I couldn't get it work before.
The key is, the AGC gain works only on the deviation of the signal from
the integrated mean. The integrated mean is updated through the
integrator-highpass stage, which I decided to throw away. Since I don't
have access to the integrated mean from the Intan, I will need to
implement this stage after all.

This required adding in extra memory items in W1 - for holding the
integrated mean; and extra memory items in A1, for holding the
integrator-highpass stage's constants.

There are some really tricky fixed-point math in here that it took me a
LOOOOONG time to understand (also documented in comments).

1. The code where the AGC-gain was updated, it was loaded into the
accumulator by multiplying by 0x4000 (16384). The after subtracting
either 0x8000 or 0x7fff from it, was extracted to half-reg with s2rnd.
The key here is that the AGC-gain is Q8.8. The value of 0x8000 and
0x7fff represents whether the gained-sample is smaller or greater than
the AGC-target. This whole procedure then becomes the equivalent of
either adding 2^-8 or subtracting 2^-7 from the Q8.8 value of the
AGC-gain.

This can be demonstrated by the following lines using functions defined
in AGC_sim.py:

gain=52
Q8_gain = int(decimal_to_Q8_8(52),16)
acc_val = int(decimal_to_Q1_31(hex_to_Q15(Q8_gain) * hex_to_Q15(16384)),16)
final_gain = mult_acc_to_halfReg(acc_val, 0, 0, 'sub', 's2rnd')
final_gain = hex_to_Q8_8(final_gain)

The arbitrary gain (52) is initially represented in Q8.8 format, which
in binary is as Q8_gain. Then multiplying it with 0x4000 in the default
mode, then extracting it to half-reg with s2rnd, the final_gain still
retains the same Q8.8 value.

2. The fact that AGC-gain only works on the signal's deviation from the
integrated mean, justifies why in gtkclient, the incoming samples were
treated as SIGNED CHAR. This is so Tim can plot the samples around 0,
which in this case is 0x0, and ranges from [-1,1]=[0x80,0x7f]. This is
in contrast with me streaming raw binary-offset samples from Intan,
where 0 is 0x8000, and [-1,1]=[0,0xff]. This then required me to change
inside gtkclient how things are displayed based on whether a signed or
unsigned sample is shown.

3. Even despite the above point, plug and chugging in the
integrator-high pass code along with AGC code does not make my system
work. I think the problem is thus:

The first step in the integrator stage is to conver the unsigned offset
binary sample from the ADC to signed Q1.15, by multiplying the sample by
1 (0x7fff) -- the result in the accumulator then is signed. However,
since the incoming samples is at most 14 bits (after the pregain), this
process will never flip it from positive to negative. Thus the s2rnd
following it simply further amplifies the sample.

For my case, on the other hand, the incoming samples are already 16
bits. Furthermore, the baseline is 0x8000. Thus doing the same
conversion to Q1.15 will result in any signal value less than the
reference level to become negative, with incorrect amplitude.

There are two solutions to this problem:
i. Simply shift the incoming samples to 12 bits. This can be done
easily.
ii. Configure Intan to send in SIGNED samples. This way, the same code
can be modified to use (default) mode instead and should work still. --
I will try this next.

4. Finally, in comparison with the streaming samples monitored on
gtkclient on Tim's headstage, my raw samples seem to have a lot of
overlaps...this is strange as I do not ever see this when tested on
JTAG. Need to try to resolve this...