AGC now works! After changing the Intan to give signed 16-bit samples. The integrator then used the default mode of multiply-accumulate rather than s2rnd, after changing the coefficients to be:
0x7fff, 0x8000, 0x7ffff, 1600; rather than previously
0x7fff, 0xc000, 0x4000, 800;

These coefficients occupy the first two 32-bit words of every frame in
A1.

This keeps the same analytical behavior of the integrator high-pass
stage. The actual AGC code didn't require any changes.

However, as Tim confirms, there still seem to be data corruption/value
warping, demonstrated by the phase lag and incomplete overlap of the
sinusoid in the spike view of the gtkclient.

I will try the 12-bit unsigned approach to see if this solves the
problem..