Edited how gtkclient reads configurations. gtkclient encodes what configuration to read in state.bin (by it's aboslute address). If we move the configuration.bin file to the directory for a different copy of gtkclient, it will probably be read, assuming its name is configuration.bin, and the existing state.bin doesn't point to some other file for its configurations.
I think this is confusing and not very user friendly. Therefore, I'm
changing it so that all configurations are read from
"configuration.bin". If that file is not found, a new
"configuration.bin" file will be created. To use previously sorted
templates, simply copy over that gtkclient's configuration.bin file and
rename the old one.

I'm also going to remove the "Set configuration as" option to make it
more consistent. Having less choice makes it easy for people to learn
the system.