Changed gtkclient so that rasters display all 128 channels of a headstage instead of just 97 (stupid UI decision in the first place!).
A weird UI quirk...under the filter setting, pressing either "osc" or
"[500,9000Hz]" will change the IIR settings of the current 4 channels to
the chosen setting.

Changing channels after that DOES NOT apply the chosen setting..the
button needs to be clicked again to apply that. Of course, by default it
is the [500,9000Hz] setting.

Also, "sync headstage" button resets the onboard IIR settings, but does
not reset the radio-button appearance. Not sure what's the best way to
do this in terms of UI.